Move isEmoji into CharWidth.ts. Rename IInputEvent -> IDOMInputEvent
