+++ src/Terminal.ts
@@ -48,12 +48,12 @@ import { ScreenDprMonitor } from './ui/ScreenDprMonitor';
 import { ITheme, IMarker, IDisposable } from 'xterm';
 import { removeTerminalFromCache } from './renderer/atlas/CharAtlasCache';
 import { DomRenderer } from './renderer/dom/DomRenderer';
-import { IKeyboardEvent, IInputEvent } from './common/Types';
+import { IKeyboardEvent, IDOMInputEvent } from './common/Types';
 import { evaluateKeyboardEvent } from './core/input/Keyboard';
 import { KeyboardResultType, ICharset } from './core/Types';
 import { clone } from './common/Clone';
 import { WebglRenderer } from './renderer/webgl/WebglRenderer';
-import { isEmoji } from './Emoji';
+import { isEmoji } from './CharWidth';
 
 // Let it work inside Node.js for automated testing purposes.
 const document = (typeof window !== 'undefined') ? window.document : null;
@@ -624,7 +624,7 @@ export class Terminal extends EventEmitter implements ITerminal, IDisposable, II
       self._keyUp(ev);
     }, true));
 
-    this.register(addDisposableDomListener(this.textarea, 'input', (ev: IInputEvent) => this._onInput(ev), true));
+    this.register(addDisposableDomListener(this.textarea, 'input', (ev: IDOMInputEvent) => this._onInput(ev), true));
     this.register(addDisposableDomListener(this.textarea, 'keydown', (ev: KeyboardEvent) => this._keyDown(ev), true));
     this.register(addDisposableDomListener(this.textarea, 'keypress', (ev: KeyboardEvent) => this._keyPress(ev), true));
     this.register(addDisposableDomListener(this.textarea, 'compositionstart', () => this._compositionHelper.compositionstart()));
@@ -1542,7 +1542,7 @@ export class Terminal extends EventEmitter implements ITerminal, IDisposable, II
     }
   }
 
-  protected _onInput(event: IInputEvent) {
+  protected _onInput(event: IDOMInputEvent) {
     if (event.inputType === 'insertText' && isEmoji(event.data)) {
       this.showCursor();
       this.handler(event.data);
